+++ test/end-to-end/generalerror.test.js
@@ -28,11 +28,14 @@ import {
     setInputValues,
     submitFormReturnRequestAndResponse,
     getGeneralError,
+    getInputAdornmentsSuccess,
+    getInputAdornmentsError,
+    getFieldErrors,
 } from "../helpers";
 
 // Run the tests in a DOM environment.
 require("jsdom-global")();
-import { TEST_SERVER_BASE_URL, SIGN_UP_API } from "../constants";
+import { TEST_SERVER_BASE_URL, SIGN_UP_API, SIGN_IN_API } from "../constants";
 
 describe("General error rendering", function () {
     before(async function () {
@@ -89,6 +92,45 @@ describe("General error rendering", function () {
 
             const generalError = await getGeneralError(page);
             assert.strictEqual(generalError, response.message);
+        });
+
+        it("Sign in error", async function () {
+            await page.evaluate(() =>
+                localStorage.setItem("SHOW_GENERAL_ERROR", "EMAIL_PASSWORD EMAIL_PASSWORD_SIGN_IN")
+            );
+            // Set values.
+            await setInputValues(page, [
+                { name: "email", value: "john.doe@supertokens.io" },
+                { name: "password", value: "Str0ngP@ssw0rd" },
+            ]);
+
+            let [{ response }] = await Promise.all([submitFormReturnRequestAndResponse(page, SIGN_IN_API)]);
+
+            assert(response.status === "GENERAL_ERROR" && response.message === "general error from API sign in");
+
+            const generalError = await getGeneralError(page);
+            assert.strictEqual(generalError, response.message);
+        });
+
+        it("Email exists error", async function () {
+            await toggleSignInSignUp(page);
+            await page.evaluate(() => localStorage.setItem("SHOW_GENERAL_ERROR", "EMAIL_PASSWORD EMAIL_EXISTS"));
+            // Set values.
+            await setInputValues(page, [
+                { name: "email", value: "john.doe@supertokens.io" },
+                { name: "password", value: "Str0ngP@ssw0rd" },
+                { name: "name", value: "John Doe" },
+                { name: "age", value: "20" },
+            ]);
+
+            const successAdornments = await getInputAdornmentsSuccess(page);
+            assert.strictEqual(successAdornments.length, 3);
+
+            const errorAdornments = await getInputAdornmentsError(page);
+            assert.strictEqual(errorAdornments.length, 1);
+            let fieldErrors = await getFieldErrors(page);
+
+            assert.deepStrictEqual(fieldErrors, ["general error from API email exists"]);
         });
     });