+++ test/server/index.js
@@ -272,6 +272,16 @@ function initST({ passwordlessConfig } = {}) {
                 apis: (oI) => {
                     return {
                         ...oI,
+                        emailExistsGET: async function (input) {
+                            let generalError = input.options.req.getKeyValueFromQuery("generalError");
+                            if (generalError === "true") {
+                                return {
+                                    status: "GENERAL_ERROR",
+                                    message: "general error from API email exists",
+                                };
+                            }
+                            return oI.emailExistsGET(input);
+                        },
                         signUpPOST: async function (input) {
                             let body = await input.options.req.getJSONBody();
                             if (body.generalError === true) {
@@ -281,6 +291,16 @@ function initST({ passwordlessConfig } = {}) {
                                 };
                             }
                             return oI.signUpPOST(input);
+                        },
+                        signInPOST: async function (input) {
+                            let body = await input.options.req.getJSONBody();
+                            if (body.generalError === true) {
+                                return {
+                                    status: "GENERAL_ERROR",
+                                    message: "general error from API sign in",
+                                };
+                            }
+                            return oI.signInPOST(input);
                         },
                     };
                 },